moving new models to seperate files
