+++ src/models/mod.rs
@@ -1,5 +1,7 @@
+pub mod args;
 pub mod card;
 pub mod card_types;
 pub mod choice;
+pub mod file_type;
 pub mod stateful_list;
 pub mod user_answer;
